Add icons and a little UI

+++ app/src/main/java/com/playground/blueskiesweather/model/DayForecast.kt
@@ -1,7 +1,7 @@
 package com.playground.blueskiesweather.model
 
-data class DayForecast(val highTemp: Double,
-                       val lowTemp: Double,
+data class DayForecast(val highTemp: Int,
+                       val lowTemp: Int,
                        val date: String,
                        val conditions: String,
                        val dayIcon: String,
